# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: ShopMe
  version: 1.0.0
  description: API for managing Italian products shop
servers:
  - url: http://localhost:3001/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        whatsappPhoneNumber:
          type: string
          nullable: true
        whatsappApiToken:
          type: string
          nullable: true
        whatsappWebhookUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        welcomeMessages:
          type: object
          additionalProperties:
            type: string
          description: Messaggio di benvenuto multilingua
        wipMessages:
          type: object
          additionalProperties:
            type: string
          description: Messaggio "Work in Progress" multilingua
        afterRegistrationMessages:
          type: object
          additionalProperties:
            type: string
          description: Messaggio post-registrazione multilingua
    CreateWorkspaceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        whatsappPhoneNumber:
          type: string
          nullable: true
        whatsappApiToken:
          type: string
          nullable: true
        whatsappWebhookUrl:
          type: string
          nullable: true
        welcomeMessages:
          type: object
          additionalProperties:
            type: string
          description: Messaggio di benvenuto multilingua
        wipMessages:
          type: object
          additionalProperties:
            type: string
          description: Messaggio "Work in Progress" multilingua
        afterRegistrationMessages:
          type: object
          additionalProperties:
            type: string
          description: Messaggio post-registrazione multilingua
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
...
    FAQ:
      type: object
      properties:
        id:
          type: string
          format: uuid
        question:
          type: string
          description: La domanda della FAQ
        answer:
          type: string
          description: La risposta alla domanda
        isActive:
          type: boolean
          description: Indica se la FAQ è attiva o disattiva
        workspaceId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateFAQRequest:
      type: object
      required:
        - question
        - answer
      properties:
        question:
          type: string
          description: La domanda della FAQ
        answer:
          type: string
          description: La risposta alla domanda
        isActive:
          type: boolean
          default: true
          description: Indica se la FAQ è attiva o disattiva
...
paths:
  /workspaces:
    get:
      summary: List all workspaces
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a new workspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /workspaces/{workspaceId}/faqs:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Lista di tutte le FAQ
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista delle FAQ per il workspace specificato
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FAQ"
        "401":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Crea una nuova FAQ
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFAQRequest"
      responses:
        "201":
          description: FAQ creata con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FAQ"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /messages:
    post:
      summary: Process a customer message
      description: >
        Processes an incoming message from a WhatsApp user and returns a response.
        Now includes language detection that will automatically detect the language
        of the message and update the customer's language preference accordingly.
        This enables the system to respond in the customer's preferred language.
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - phoneNumber
                - workspaceId
              properties:
                message:
                  type: string
                  description: The message text from the customer
                  example: "Hello, I want to know more about your products"
                phoneNumber:
                  type: string
                  description: The phone number of the sender in international format
                  example: "+391234567890"
                workspaceId:
                  type: string
                  description: The ID of the workspace to process the message in
                  example: "clj1z6qxr0000qwzlgt7a02jm"
                sessionId:
                  type: string
                  description: Optional session ID for existing conversations
                  example: "clj1z6qxr0001qwzlgt7a02jm"
                isNewConversation:
                  type: boolean
                  description: Flag to indicate this is a new conversation
                  example: false
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      processedMessage:
                        type: string
                        example: "Thank you for your message. Our best-selling product is Italian extra virgin olive oil."
                      detectedLanguage:
                        type: string
                        description: The detected language of the customer's message
                        example: "en"
                      sessionId:
                        type: string
                        example: "clj1z6qxr0001qwzlgt7a02jm"
                      customerId:
                        type: string
                        example: "clj1z6qxr0002qwzlgt7a02jm"
        '400':
          description: Bad request - invalid parameters
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...